Undo refactoring
