+++ lib/drives/index.js
@@ -1,4 +1,5 @@
 const hyperdrive = require('hyperdrive')
+const hypercoreCrypto = require('hypercore-crypto')
 const datEncoding = require('dat-encoding')
 const pump = require('pump')
 const sub = require('subleveldown')
@@ -7,7 +8,6 @@ const collectStream = require('stream-collector')
 const { NanoresourcePromise: Nanoresource } = require('nanoresource-promise/emitter')
 const { Transform } = require('streamx')
 
-
 const {
   fromHyperdriveOptions,
   fromStat,
@@ -15,6 +15,7 @@ const {
   fromMetadata,
   fromDriveConfiguration,
   fromNetworkConfiguration,
+  toNetworkConfiguration,
   toHyperdriveOptions,
   toStat,
   toMount,
@@ -33,11 +34,11 @@ const log = require('../log').child({ component: 'drive-manager' })
 const TRIE_UPDATER_SYMBOL = Symbol('hyperdrive-daemon-trie-updater')
 
 class DriveManager extends Nanoresource {
-  constructor (corestore, network, db, opts = {}) {
+  constructor (corestore, networking, db, opts = {}) {
     super()
 
     this.corestore = corestore
-    this.network = network
+    this.networking = networking
     this.db = db
     this.opts = opts
     this.watchLimit = opts.watchLimit
@@ -65,8 +66,19 @@ class DriveManager extends Nanoresource {
 
   async _open () {
     return Promise.all([
+      this._rejoin(),
       this._remirror()
     ])
+  }
+
+  async _rejoin () {
+    if (this.noAnnounce) return
+    const seedList = await dbCollect(this._seedIndex)
+    for (const { key: discoveryKey, value: networkOpts } of seedList) {
+      const opts = networkOpts && networkOpts.opts
+      if (!opts || !opts.announce) continue
+      this.networking.join(discoveryKey, { ...networkOpts.opts })
+    }
   }
 
   async _remirror () {
@@ -177,6 +189,64 @@ class DriveManager extends Nanoresource {
         return resolve()
       })
     })
+  }
+
+  async configureNetwork (discoveryKey, opts = {}) {
+    const self = this
+    const encodedKey = datEncoding.encode(discoveryKey)
+    const networkOpts = {
+      lookup: !!opts.lookup,
+      announce: !!opts.announce,
+    }
+    const seeding = opts.lookup || opts.announce
+    var networkingPromise
+
+    const sameConfig = sameNetworkConfig(discoveryKey, opts)
+    // If all the networking options are the same, exit early.
+    if (sameConfig) return
+
+    const networkConfig = { opts: networkOpts }
+    if (seeding) await this._seedIndex.put(encodedKey, networkConfig)
+    else await this._seedIndex.del(encodedKey)
+
+    // Failsafe
+    if (networkOpts.announce && this.noAnnounce) networkOpts.announce = false
+
+    try {
+      if (seeding) {
+        networkingPromise = this.networking.join(discoveryKey, networkOpts)
+      } else {
+        networkingPromise = this.networking.leave(discoveryKey)
+      }
+      networkingPromise.then(configurationSuccess)
+      networkingPromise.catch(configurationError)
+    } catch (err) {
+      configurationError(err)
+    }
+
+    function sameNetworkConfig (discoveryKey, opts = {}) {
+      const swarmStatus = self.networking.status(discoveryKey)
+      if (!swarmStatus) return opts.lookup === false && opts.announce === false
+      return swarmStatus.announce === opts.announce && swarmStatus.lookup === opts.lookup
+    }
+
+    function configurationError (err) {
+      log.error({ err, discoveryKey: encodedKey }, 'network configuration error')
+    }
+
+    function configurationSuccess () {
+      log.debug({ discoveryKey: encodedKey }, 'network configuration succeeded')
+    }
+  }
+
+  async getNetworkConfiguration (discoveryKey) {
+    const networkOpts = await dbGet(this._seedIndex, datEncoding.encode(discoveryKey))
+    return networkOpts ? networkOpts.opts : null
+  }
+
+  async getAllNetworkConfigurations () {
+    const storedConfigurations = (await dbCollect(this._seedIndex)).map(({ key, value }) => [key, value])
+    return new Map(storedConfigurations)
   }
 
   async getAllStats (opts) {
@@ -310,10 +380,10 @@ class DriveManager extends Nanoresource {
     if (opts.fuseNetwork) {
       // TODO: The Network drive does not announce or remember any settings for now.
       initialConfig = { lookup: true, announce: false, remember: false }
-      await this.network.configure(drive.metadata, initialConfig)
+      await this.configureNetwork(drive.metadata, initialConfig)
     } else if (!drive.writable || opts.seed) {
       initialConfig = { lookup: true, announce: false, remember: true }
-      await this.network.configure(drive.metadata, initialConfig)
+      await this.configureNetwork(drive.metadata, initialConfig)
     }
 
     // Make sure that any inner mounts are recorded in the drive index.
@@ -323,15 +393,15 @@ class DriveManager extends Nanoresource {
       const mountKey = feed.key.toString('hex')
 
       log.info({ discoveryKey: feed.discoveryKey.toString('hex') }, 'registering mountpoint in drive index')
-      const parentConfig = (await this.network.getConfiguration(drive.discoveryKey)) || initialConfig || {}
-      const existingMountConfig = (await this.network.getConfiguration(feed.discoveryKey)) || {}
+      const parentConfig = (await this.getNetworkConfiguration(drive.discoveryKey)) || initialConfig || {}
+      const existingMountConfig = (await this.getNetworkConfiguration(feed.discoveryKey)) || {}
       const mountConfig = {
         lookup: (existingMountConfig.lookup !== false) && (parentConfig.lookup !== false),
         announce: !!(existingMountConfig.announce || parentConfig.announce),
         remember: true
       }
 
-      if (mountConfig) await this.network.configure(feed.discoveryKey, mountConfig)
+      if (mountConfig) await this.configureNetwork(feed.discoveryKey, mountConfig)
       this.emit('configured-mount', feed.key)
       this._configuredMounts.add(mountKey)
 
@@ -412,8 +482,7 @@ class DriveManager extends Nanoresource {
   }
 
   async _rpcAllNetworkConfigurations (call) {
-    return this.network._rpcAllConfigurations(call)
-    const networkConfigurations = await this.network.getAllConfigurations()
+    const networkConfigurations = await this.getAllNetworkConfigurations()
 
     const rsp = new rpc.drive.messages.NetworkConfigurationsResponse()
     rsp.setConfigurationsList([...networkConfigurations].map(([, value]) => toNetworkConfiguration({
@@ -452,7 +521,7 @@ class DriveManager extends Nanoresource {
     const drive = this.driveForSession(id)
     const opts = fromNetworkConfiguration(call.request.getNetwork())
 
-    await this.network.configure(drive.metadata.discoveryKey, { ...opts })
+    await this.configureNetwork(drive.metadata.discoveryKey, { ...opts })
 
     const rsp = new rpc.drive.messages.ConfigureNetworkResponse()
     return rsp
